Reorganizacion de componentes
