+++ src/app/app-routing.module.ts
@@ -1,15 +1,15 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 
-import { MenuComponent } from './components/menu/menu.component';
-import { InicioComponent } from './components/inicio/inicio.component';
-import { PaguinaNoEncontradaComponent } from './components/paguina-no-encontrada/paguina-no-encontrada.component';
+import { MenuComponent } from './components/Otros/menu/menu.component';
+import { InicioComponent } from './components/Otros/inicio/inicio.component';
+import { PaguinaNoEncontradaComponent } from './components/Otros/paguina-no-encontrada/paguina-no-encontrada.component';
 import { AutenticacionGuard } from "./guard/autentificacion.guard";
 import { InventarioComponent } from './components/Inventario-All/inventario/inventario.component';
-import { CajasComponent } from './components/cajas/cajas.component';
-import { ListaDeComprasComponent } from './components/lista-de-compras/lista-de-compras.component';
-import { ClientesComponent } from './components/clientes/clientes.component';
-import { CondicionesDeUsosComponent } from './components/condiciones-de-usos/condiciones-de-usos.component'
+import { CajasComponent } from './components/Caja/cajas/cajas.component';
+import { ListaDeComprasComponent } from './components/Compras/lista-de-compras/lista-de-compras.component';
+import { ClientesComponent } from './components/Cliente/clientes/clientes.component';
+import { CondicionesDeUsosComponent } from './components/Otros/condiciones-de-usos/condiciones-de-usos.component'
 
 const routes: Routes = [
   {
